[KAIZEN-0] fjerne runner-konsept
ikke behov for ekstra isolasjon her
